pyproject.toml
[project]
name = "cronical"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cron-descriptor>=2.0.8",
    "cronitor>=4.9.0",
    "python-crontab>=3.3.0",
    "python-dotenv>=1.2.2",
]
